canonizacion-requirements.txt
attrs==22.1.0
click==8.1.3
coverage==6.5.0
Flask==2.2.2
Flask-SQLAlchemy==3.0.2
flask-swagger==0.2.14
iniconfig==1.1.1
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
packaging==21.3
pluggy==1.0.0
py==1.11.0
pyparsing==3.0.9
pytest==7.1.3
PyYAML==6.0.1
SQLAlchemy==1.4.44
tomli==2.0.1
Werkzeug==2.2.2
pulsar-client==3.5.0
pulsar-client[avro]==3.5.0
psycopg2==2.9.9
greenlet==3.1.1
PyDispatcher==2.0.6
